Refactor code (well not really just renamed already refactored code to fit the new standard)
