+++ data/task_lists.txt
@@ -6,3 +6,5 @@ E | 0 | lol | 2019-10-15 | 2019-10-15
 E | 1 | lolwdad | 2019-10-15 | 2019-10-15
 E | 0 | lol | 2018-07-13 | 2026-05-23
 T | 0 | ok | NA | NA
+E | 0 | wdwadwadad | 2001-03-03 | 2004-03-03
+T | 0 | ok | NA | NA
